requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.9
djangorestframework==3.12.4
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
